requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.6b0
bootstrap-admin==0.4.4
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.2.4
django-admin-tools==0.9.1
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
django-debug-toolbar-request-history==0.1.3
django-extensions==3.1.3
django-filter==2.4.0
django-material==1.9.0
django-mathfilters==1.0.0
django-nested-inline==0.4.3
django-rest-swagger==2.2.0
django-reverse-admin==2.9.6
django-reversion==3.0.9
djangorestframework==3.12.4
docopt==0.6.2
drf-nested-routers==0.93.3
drf-yasg==1.20.0
et-xmlfile==1.1.0
html5lib==1.1
inflection==0.5.1
itypes==1.2.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==3.0.1
jsonfield==3.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
mypy-extensions==0.4.3
mysqlclient==2.0.3
openapi-codec==1.3.2
openpyxl==3.0.7
packaging==20.9
parso==0.8.2
pathspec==0.8.1
Pillow==8.2.0
prompt-toolkit==3.0.19
ptpython==3.0.17
pycparser==2.20
pydot==1.4.2
Pygments==2.9.0
pyparsing==2.4.7
Pyphen==0.10.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
tinycss2==1.1.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.5
wcwidth==0.2.5
WeasyPrint==52.5
webencodings==0.5.1
pdfkit